changed cmp to use fzf for buffers
remaped <c-e> to <a-e> so you can get character form line below. And
still use <c-e> to escape cmp. Also now matches <a-y>.

I didn't add the fuzzy find for the path. It is not quite as useful and
it required the command line program fd. Which I might not have on every
computer.