attempt to fix ctrl-v as paste
I've atempted to fix ctrl-v so that it pasts without copying first. Like
it does in regualr editors. however I cannot test to see if it works
currently because I'm using ubuntu within win11 wsl. And I havne't got
the clipboard working here. I will test this in the next couple of days.
(but it hasn't borken anything)