vhdl_depends.lua inital commit
I've written a lua script to find VHDL dependices and write them out in
a list. This has been written with some limitations. It only searches
for direct instantation within the work library.

It also assumes that the name you call the file matches the  dependcies'
name.

I will also need to add a dependicey file support for other library
searching and conflict resolution.

But it is a good start and should satify 90% of my issues.